Explicitly disable Paragraph Layout support in bundled ICU
The bundled copy of ICU doesn't have the code for this optional feature,
even if we find the icu-le-hb library during configure. In versions of
mozjs52 compiled before Debian had icu-le-hb available, we didn't try
to build that feature, so not having it is not a regression.

Patch from Mike Hommey, via mozilla-unified.

Closes: #902197