debian/gbp.conf: Repack the mozjs sources using official tooling
Filter out even more files from the repo using the gbp postunpack command,
replacing the content of the archive with the needed one as per the official
mozilla tarball. This ensures also that we don't filter anything we need.

Unfortunately we can't remove the same files from the pristine-tar committed
archive as the scripts runs after this.

It would probably worth to improve gbp to handle this case though.