d/rules: Swap --host and --target
The Mozilla build system's terminology is not the same as Autoconf's,
and previous versions of the mozjs Debian packages got the translation
wrong.

The correct terminology appears to be:

                              Autoconf/dpkg       Mozilla
    machine we build on         build               host
    machine we build for        host               target

For example, in the common case of using an x86 machine to build ARM
binaries, dpkg and Autoconf would say we have an x86 build and an ARM
host, whereas Mozilla would say we have an x86 host and an ARM target.